progress is moving slow. I have added some detail to the JSON and debugged minor problems.
Your desiging on this site is lacking
Please come up with a better way to
implement the 'ThreePaneInfoPanel' and
'ContentBox' which currently enables the
neat browsing of the skills page.

The actual display of info could be handled better, essentially you
you have sub components on the page that
the user can route to with the nav bar.
each component would break down the lowest
branch on the JSON tree.

Then skill areas are still collapsable,
but the page has FLOW