All I did was clean up the json. There was some really nasty unprofessional stuff.  kinda sucks being this is github so even erasing it now doesn't really get rid of it. DRAT
Well. Im getting it cleaned up.
like i mentioned previously,
the casing structure is still going
to stay a little odd, only reason being
I want to keep the titles in the correct case,
so I dont have to use css to manually do this
otherwise the file struct is mostly conventional meaning
if there is a multi word thing it should be
```camelCase```