Got experience working. Doesnt look pretty but solved the persistent bug of not being able to view experiences where multiple positions were worked. our solution has limitations, noted in doc string for MultiListRender.
basically needed to make the variable/ arg
props.duties into an array of objects where
key is position title and value is array of
strings for duties worked in that position.
This was the orignal format in JSON but
something got lost in transit it seems. ONTO
BIGGER AND BETTER THINGS