Very slow progress today. Switched to Sass for styling, want to make the site more responsive before putting it on live site.
got hung up on installing neovim
so we can do non-IDE search and replace
with vim: ```:%s/old/new/g```
and also messed around with making
fast and easy key bindings, this wasted a lot of time
hoping we can get back tonight to get
the styling responsive and condensed.
if not,. maybe time to switch gears back
to angular or python so we can be slightly more producitive.