added the css styles on panel content from a previous commit (before we ditched all our styles) to get the css impl of info panel working.
like noted, wed like to get this to work just with react,
so css styles are not inherently tied to UI functioning
properly ( graceful degradation/ progressive enhancement).

identified some todos and other general problems
with the navbar impl, chiefly that would be
the .content div which supposed to seperate the nav
bar from the page content/layout is overflowing
(gives the page a nice background) however, we
want to keep this background but have child
elements appear in normal flow (directly below
navbar) its tricky.