Addressed the two big issues mentioned last time: the view port height being too large and overflowing, and the navbar last item being snagged to side of screen.
we used grid to change the item placement to start center
instead of center, and decreased the view port height.
sites looking pretty good.

Added a intro paragraph for each sub section of skills,
some typography styling, and an animated box shadow
for the "examine" button.

in process of including the intro text for each sub area,
created a sort of silly bug since that text is deleted manually,
toggling the window open and then closed means
the intro text is no longer rendered. Felt like that
delete syntax looked jenky. There should be
an obvious alternative, which we will be working
on next time