We are sitting pretty now. fixed 3 major issues in skills and about. 1. we were setting the state of slider with the value of its props, which was unadvisable and causing this component to not register changes when its props were modified. Instead, it just reads the value from props and now its displayed value is updated on page chages.
2. moved elements around in closed vs open row,
so that intro text is still displayed when row is open,
and the same button is used for opening and closing,
while the text inside changes. the content
when the row is open now appears below the heading
and intro body text.

3. moved skills and about info panels into their own rfc in
app component, which means the row dropdowns
are closed when we change the url or refresh the page.

At this point, once we get the projects page and home page
finished, all we need to do is finish styling and making the
site responsive, then get on with hosting it and finally
we can move attention back to more interesting projects.