Smoll debugging and css problem solving/fine tuning.
We first tried to use media queries
to show .dd-col hidden content
on mobile view using the
:active pseduo-class, but this
causes it to appear in a blink of the eye
(only when the mouse is clicked DOWN)

Instead, we chose to just show everything for now
when the user has clicked a topic to examine.
Doesnt look wonderful, but it does work decently
well now.

Tried styling the range/slider typed input element
used to display numeric data in the info panel,
again, got it styled, doesnt look great, but it works
for now. What I wanted to accomplish is different-
a colored background for the active part of the range
(sort of like a loading percent bar)

Thoughts on resolving this is avoid using
the input[type=range] altogether, and accomplish
this with a nested span and the css calc() function
to fill only a percent of the parent span.

While adjusting description data for projects.jsx,
I realized what we should have done from the start
was use the github api to show this data,
I'm happy to start a branch now where we re do
the projects component with the github api, while
also trying to implement a popular component library
in the process - semantic UI (why not kill two birds
with one stone?) the current source code
for projects page looks terrible anyway so no harm no foul.