Ditched the Awad components since we dont have any use for them.
Minor edits of the json, making labels/heading more
concise and clear.

we experimented with different implementations
of the range so that the value is shown
as a tooltip when hovering over a range,
no luck. Instead, chose to create our own
range/slider, which looks better and will
support animation with ease. The only
draw back is that we needed to use
inline style to create this slider.