Minor responsiveness fine tuning: adjusted the headings for each info col so that they dont cram on smaller screen sizes and become unaligned, centered the custom slider.
Tried implementing a 'close all' button so
users can unexpand multiple columns rapidly.
its a 'nice to have' feature so we commented it
out after struggling to come up with an idea for
how to implement it.

Made the dropdown columns hidden by default,
and their display is triggered by clicking
this increases mobile friendliness (no hover on cellphone)