Final touches. added CSS responsiveness, ScrollToTop Effect ( does what it sounds like). Added the ability to show images and all the images we will need
The only real thing remaining is giving users the ability
to click on an image and enlarge it.

I am thinking we may want to just create a whole new
componentry for this, as implmenting it ontop
of the existing info panel would be SUCH a mess.
(already have imageViewer, try that?)

after that, next steps are final proof read, then
DEPLOYMENT (AT LAST!!!!)