Instead of creating a seperate branch, made a new component called 'panel' which currently works like the 'browsePhotos' component in OralHistory project, either displaying the minified images or an animated modal that enlarges them
neat detail is that when the screen gets smaller than
600px, it switches from the vertical layout to
a refactored RFC ImageSlider, so users can swipe
thru pictures rather than scrolling like they have to do
constantly with this site.

Next steps are design a layout and then replace the
POC with a refactored, better organized version
of info panel. Put great thought into how data is
organized when it gets to the dropdown column
level, because that is when the imagePicker or
ImageSlider will get them.