Dealt with (some of) the organization concerns by breaking Panel.jsx into two files ProgressBlog and PhotoGallery. This pretty much deals with all my bloat complaints in Panel.jsx, except now the hard coding of blog posts is in ProgressBlog.jsx instead.
Created some dumb components and manager for
blog post details- links, images and code block can
now be displayed in details, assuming the detail item
is an object with a paticular propety (not the safest way
of deciding what to render)

hoping to deal with connecting this project to firebase and
firestore today, as well as some general todos. Another
quality of life concern crops up now that we dealt with
the bloat, which is passing the json items far down into
the component tree. Hoping when we migrate our datastores
to firestore this will be handled/ designed out.