Finished configuring firestore, and we are reading documents from the progressBlog.
however, progress blog is down at this point.
not sure yet what the issue is but we are pressed for
time atm and cannot debug it. Only clue is that
the configuration and security rules are exactly the same
as other projects, but more secure - api key has not been
publicized and rules allow reading but not writing for any
unauthorized user. yet the hook that reads the 'progressBlog'
collection returns an empty array, with no errors, despite
there being documents in the collection with the correct format.