just testing if we are doing the wrong thing by removing REACT_APP in the .env file
Found a stack overflow answer, suggesting that
we should write a script, invoked at build time,
that requires the .env file and defines the env variables.
because netlify doesnt look at the env file prior to build,
meaning if the deletion done in this commit was not
progress, it still has no effect on the deployed application.