We've done a lot of optimizing today, got rid of a lot of nagging warnings. Fixed some inconsistencies in the way pages are displayed, addressed some issues with mobile landscape view.
Generally, Id say we need to put a little more
thought towards the way we are structuring
styles. Theres a lot of sub points to address here,
suggest you make a todo list issue and link that
to a blog post tonight. As well as dealing with this,
hope to put getDocs in its own useEffect hook,
and address some older todos, like making more
of the site content loaded dynamically, and adding
links to resume pdf and linked in to navbar or homepage.

Honestly theres a lot todo on this site. Feels good to
get this much work done, but we need to reorganize
or else progress is painfully inefficient (styles especially,
mediaquery or not is counter intutitive, fix one thing, break
something else) and this will take forever.