Nice big refactor of Experience component, got this work experience data on firebase.
loaded this data, using grid we now
have a two pain smart / dumb component
window display, where companies ive worked for
are displayed on left and when a selection is made,
job info and responsiblities are shown on right.

Added a timeout for loading screens across the
site, so these dont flash for a milisecond, in cases
where user has already navigated to the page once.
need to make the fns that call firestore async, so
we dont have cases where timeout finishes, but
documents havent been retrieved from snapshot.

fixed some quality of life / responsiveness things
across the site, such as margin issue with avatar
img on homepage.